+++ promql/functions_test.go
@@ -51,7 +51,7 @@ func TestDeriv(t *testing.T) {
 	// https://github.com/prometheus/prometheus/issues/7180
 	for i = 0; i < 15; i++ {
 		jitter := 12 * i % 2
-		a.Append(0, metric, int64(start+interval*i+jitter), 1)
+		a.Append(0, metric, start+interval*i+jitter, 1)
 	}
 
 	require.NoError(t, a.Commit())
